refactor: extract TreeNode out of file-tree-view
